requirements.txt
Django==4.2.10
python-telegram-bot==20.7
openai==1.12.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
django-environ==0.11.2
django-allauth==0.61.1
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
Pillow==10.2.0
django-ckeditor==6.7.1
django-celery-beat==2.5.0
celery==5.3.6
redis==5.0.1 